Cleanup. TODO: fix tests
